Get table info and recursively search wiki links
The spider now yields some basic table info, extracting the
unadorned text from the header and data of each html table row.
The title of the article is also returned.

This spider recursively searches wiki links found in the body of the
response as well, up to a limit determined by the spider definition.
Currrently that limit is 1, so it only follows direct links in the
first page.